\name{GDPgrowth}
\alias{GDPgrowth}
\title{Panel of GDP growth rates of 26 EU countries}
\description{Panel data of GDP growth rates of 26 European Union countries from 1999 to 2011. Countries included
             are: Austria, Belgium, Bulgaria, Cyprus, The Czech Republic, Denmark, Estonia, Finland, France,
             Germany, Greece, Hungary, Ireland, Italy, Latvia, Lithuania, Luxemburg, Netherlands, Poland,
             Portugal, Romania, Slovak Republic, Slovenia, Spain, Sweden, The United Kingdom. Number of sections
             is 26 and number of years is 13, which brings total number of observations to 338 observations.}
\format{A tab file in the form indicated for input file (Table1 in manual).}
\source{Penn World Table 9.0
        DOI: 10.15141/S5J01T
        URL: http://www.rug.nl/ggdc/productivity/pwt/
        Feenstra, Robert C., Robert Inklaar and Marcel P. Timmer (2015), "The Next Generation of the Penn World          Table" American Economic Review, 105(10), 3150-3182
}
