\name{RGDPpc}
\alias{RGDPpc}
\title{Panel of real GDP per capita for 154 countries}
\description{Panel data of real GDP per capita for 154 countries from 1971 to 2014. Number of sections is 154 and number of years is 44, which brings total number of observations to 6776 observations.}

\format{A tab file in the form indicated for input file (Table 1 in manual).}
\source{Penn World Table 9.0
  DOI: 10.15141/S5J01T
  URL: http://www.rug.nl/ggdc/productivity/pwt/
    Feenstra, Robert C., Robert Inklaar and Marcel P. Timmer (2015), "The Next Generation of the Penn World Table" American Economic Review, 105(10), 3150-3182
}
